Fix selenium race conditions. Fix #15

+++ ckanext/oauth2/tests/test_selenium.py
@@ -20,23 +20,31 @@
 
 import unittest
 import os
-import time
-
-from urlparse import urljoin
-from nose_parameterized import parameterized
-from selenium import webdriver
 from subprocess import Popen
+import time
+from urlparse import urljoin
+
+from parameterized import parameterized
+from selenium import webdriver
+from selenium.webdriver.common.by import By
+from selenium.webdriver.support import expected_conditions as EC
+from selenium.webdriver.support.ui import WebDriverWait
 
 IDM_URL = "https://account.lab.fiware.org"
 FILAB2_MAIL = "filab2@mailinator.com"
 FILAB3_MAIL = "filab3@mailinator.com"
 FILAB_PASSWORD = "filab1234"
+PASS_INTEGRATION_TESTS = os.environ.get("INTEGRATION_TEST", "").strip().lower() in ('1', 'true', 'on')
 
-
+@unittest.skipUnless(PASS_INTEGRATION_TESTS, "set INTEGRATION_TEST environment variable (e.g. INTEGRATION_TEST=true) for running the integration tests")
 class IntegrationTest(unittest.TestCase):
 
     @classmethod
     def setUpClass(cls):
+        # nose calls this method also if they are going to be skiped
+        if not PASS_INTEGRATION_TESTS:
+            return
+
         env = os.environ.copy()
         env['DEBUG'] = 'True'
         env['OAUTHLIB_INSECURE_TRANSPORT'] = 'True'
@@ -44,37 +52,43 @@ class IntegrationTest(unittest.TestCase):
 
         cls.driver = webdriver.Firefox()
         cls.base_url = 'http://localhost:5000/'
-        cls.driver.implicitly_wait(5)
         cls.driver.set_window_size(1024, 768)
 
         cls.driver.get(IDM_URL)
         cls._introduce_log_in_parameters()
         cls.driver.get(urljoin(IDM_URL, '/idm/myApplications/361020fd7cf64456890dd98da88e64f3/edit/'))
-        cls.driver.find_element_by_id("id_callbackurl").clear()
-        cls.driver.find_element_by_id("id_callbackurl").send_keys(urljoin(cls.base_url, '/oauth2/callback'))
+        id_callbackurl = WebDriverWait(cls.driver, 10).until(EC.presence_of_element_located((By.ID, "id_callbackurl")))
+        id_callbackurl.clear()
+        id_callbackurl.send_keys(urljoin(cls.base_url, '/oauth2/callback'))
         cls.driver.find_element_by_xpath("//button[@type='submit']").click()
 
     @classmethod
     def tearDownClass(cls):
+        # nose calls this method also if they are going to be skiped
+        if not PASS_INTEGRATION_TESTS:
+            return
+
         cls._process.terminate()
         cls.driver.quit()
 
     @classmethod
     def _introduce_log_in_parameters(cls, username=FILAB2_MAIL, password=FILAB_PASSWORD):
         driver = cls.driver
-        driver.find_element_by_id("id_username").clear()
-        driver.find_element_by_id("id_username").send_keys(username)
+        id_username = WebDriverWait(cls.driver, 10).until(EC.presence_of_element_located((By.ID, "id_username")))
+        id_username.clear()
+        id_username.send_keys(username)
         driver.find_element_by_id("id_password").clear()
         driver.find_element_by_id("id_password").send_keys(password)
         driver.find_element_by_xpath("//button[@type='submit']").click()
+        WebDriverWait(driver, 30).until(EC.staleness_of(id_username))
 
     def delete_cookies(self, domain):
         self.driver.get(domain)
         self.driver.delete_all_cookies()
 
     def setUp(self):
-        self.delete_cookies(IDM_URL)
         self.delete_cookies(self.base_url)
+        self.delete_cookies(IDM_URL)
         self.verificationErrors = []
         self.accept_next_alert = True
 
@@ -84,20 +98,22 @@ class IntegrationTest(unittest.TestCase):
     def _log_in(self, referer, username=FILAB2_MAIL, password=FILAB_PASSWORD, authorize=True):
         driver = self.driver
         driver.get(referer)
-        driver.find_element_by_link_text("Log in").click()
+        WebDriverWait(driver, 30).until(lambda driver: driver.current_url == referer)
+
+        WebDriverWait(driver, 30).until(EC.element_to_be_clickable((By.LINK_TEXT, "Log in"))).click()
         self._introduce_log_in_parameters(username, password)
 
         if driver.current_url.startswith(IDM_URL) and authorize:
-            driver.find_element_by_xpath("//button[@type='submit']").click()
+            WebDriverWait(driver, 30).until(EC.element_to_be_clickable((By.XPATH, "//button[@type='submit']"))).click()
 
     def test_basic_login(self):
         driver = self.driver
         self._log_in(self.base_url)
+        WebDriverWait(driver, 20).until(lambda driver: (self.base_url + 'dashboard') == driver.current_url)
         self.assertEqual("filab2 Example User", driver.find_element_by_link_text("filab2 Example User").text)
-        self.assertEqual(self.base_url + 'dashboard', driver.current_url)
         driver.find_element_by_link_text("About").click()
+        WebDriverWait(driver, 20).until(lambda driver: (self.base_url + 'about') == driver.current_url)
         self.assertEqual("filab2 Example User", driver.find_element_by_css_selector("span.username").text)
-        self.assertEqual(self.base_url + "about", driver.current_url)
         driver.find_element_by_css_selector("a[title=\"Edit settings\"]").click()
         time.sleep(3)   # Wait the OAuth2 Server to return the page
         assert driver.current_url.startswith(IDM_URL + "/settings")
@@ -105,28 +121,27 @@ class IntegrationTest(unittest.TestCase):
     def test_basic_login_different_referer(self):
         driver = self.driver
         self._log_in(self.base_url + "about")
+        WebDriverWait(driver, 20).until(lambda driver: (self.base_url + 'about') == driver.current_url)
         self.assertEqual("filab2 Example User", driver.find_element_by_css_selector("span.username").text)
-        self.assertEqual(self.base_url + "about", driver.current_url)
         driver.find_element_by_link_text("Datasets").click()
+        WebDriverWait(driver, 20).until(lambda driver: (self.base_url + 'dataset') == driver.current_url)
         self.assertEqual("filab2 Example User", driver.find_element_by_css_selector("span.username").text)
-        self.assertEqual(self.base_url + "dataset", driver.current_url)
 
     def test_user_access_unauthorized_page(self):
         driver = self.driver
         self._log_in(self.base_url)
-        time.sleep(3)   # Wait until the log in proccess is completed
         driver.get(self.base_url + "ckan-admin")
 
         # Check that the user has been redirected to the main page
-        self.assertEquals(self.base_url, driver.current_url)
+        WebDriverWait(driver, 10).until(lambda driver: driver.current_url == self.base_url)
         # Check that an error message is shown
         assert driver.find_element_by_xpath("//div/div/div/div").text.startswith("Need to be system administrator to administer")
 
     def test_register_btn(self):
         driver = self.driver
         driver.get(self.base_url)
-        driver.find_element_by_link_text("Register").click()
-        self.assertEqual(IDM_URL + "/sign_up/", driver.current_url)
+        WebDriverWait(driver, 10).until(EC.element_to_be_clickable((By.LINK_TEXT, "Register"))).click()
+        WebDriverWait(driver, 10).until(lambda driver: driver.current_url == (IDM_URL + "/sign_up/"))
 
     @parameterized.expand([
         ("user/register", IDM_URL + "/sign_up/"),
@@ -135,8 +150,7 @@ class IntegrationTest(unittest.TestCase):
     def test_register(self, action, expected_url):
         driver = self.driver
         driver.get(self.base_url + action)
-        time.sleep(3)   # Wait the OAuth2 Server to return the page
-        self.assertEqual(expected_url, driver.current_url)
+        WebDriverWait(driver, 10).until(lambda driver: driver.current_url == expected_url)
 
 
 if __name__ == "__main__":
